+++ tests/testthat/test-maps.R
@@ -24,9 +24,10 @@ test_that("SpecimenMap R6 Subclass", {
 
   # Border Features ------------------------------------------------------------
   expect_type(vouchers$features, type = "closure")
-  voucher_features <- vouchers$features(sf_border = "black")
+  expect_error(vouchers$features(sf_border = "black", sf_states = list()))
 
   # County / State layer simple features
+  voucher_features <- vouchers$features(sf_border = "black")
   purrr::walk(
     .x = list(
       voucher_features[[1]][[1]],
@@ -123,6 +124,30 @@ test_that("SpecimenMap R6 Subclass", {
   expect_identical(
     class(voucher_repel),
     expected = c("LayerInstance", "Layer", "ggproto", "gg")
+  )
+
+  # Optionally set specific specimen records in addition to collection search.
+  voucher_labels <-
+    tibble::tribble(
+      ~"Collector", ~"Collection_Number", ~"Longitude", ~"Latitude",
+      "A. Gray", 5, 46, -111
+    )
+  expect_identical(
+    dplyr::select(vouchers$repel(vouchers = voucher_labels)$data, -"label"),
+    voucher_labels
+  )
+  expect_identical(
+    vouchers$repel(vouchers = voucher_labels)$data$label,
+    expected = "Gray\n5"
+  )
+  expect_equal(
+    nrow(
+      vouchers$repel(
+        "Hooker" = 1:4,
+        vouchers = voucher_labels
+      )$data
+    ),
+    expected = 5
   )
 
   # Map Plots ------------------------------------------------------------------